# Reduce TextFormatter allocations