--- a/Benchmarks/Text/TextFormatter/ReplaceCRLFWithSpace.cs
+++ b/Benchmarks/Text/TextFormatter/ReplaceCRLFWithSpace.cs
@@ -8,6 +8,7 @@ namespace Terminal.Gui.Benchmarks.Text.TextFormatter;
 /// Benchmarks for <see cref="Tui.TextFormatter.ReplaceCRLFWithSpace"/> performance fine-tuning.
 /// </summary>
 [MemoryDiagnoser]
+[BenchmarkCategory (nameof (Tui.TextFormatter))]
 public class ReplaceCRLFWithSpace
 {