--- a/Terminal.Gui/Text/TextFormatter.cs
+++ b/Terminal.Gui/Text/TextFormatter.cs
@@ -2444,7 +2444,7 @@ public class TextFormatter
             return text;
         }
 
-        const int maxStackallocCharBufferSize = 512; // ~1 kiB
+        const int maxStackallocCharBufferSize = 512; // ~1 kB
         char[]? rentedBufferArray = null;
         try
         {