… Benchmarks/Text/TextFormatter/RemoveHotKeySpecifier.cs